identity: fix D1 bigint bind panic, retire hostos_mcp knock auto-mint for everyone
Two fixes to the previous commit, both found by exercising it live
against production, not by review:

1. D1_TYPE_ERROR: Type 'bigint' not supported. worker's From<i64> for
   JsValue produces a JS BigInt; D1's bind rejects those outright.
   h_approve_knock already carries this exact fix and comment --
   auth_users.rs's three i64 timestamp binds (session/verification
   expiry, user created_at) didn't, and registration 500'd on the
   first real call. Cast through f64, same as the existing code.

2. Founder correction: gating hostos_mcp's knock-approval auto-mint by
   comparing this instance's owner_id against CURATOM_OWNER_ID (this
   commit's first attempt) was the wrong shape. The real plan is a
   general "connect your own MCP server" dashboard feature for every
   account, hostos-mcp included as one more entry a person adds --
   not a resource special-cased into the knock/approval path by name,
   for anyone, founder included. That branch in h_approve_knock is
   retired outright: approving a hostos_mcp knock now returns 410
   instead of minting a capabilities row through any path. The one
   live capability row this session had minted earlier was revoked
   directly against production D1.

Verified live end to end after both fixes: POST /auth/register really
sent a ZeptoMail verification email, and (verified by count, not
assumed -- the pending_verifications row was already gone) a real
account exists in D1 by the time this was checked.

+++ workers/api/src/auth_users.rs
@@ -152,8 +152,12 @@ async fn create_session(env: &Env, user_id: &str) -> Result<String> {
     .bind(&[
         hash.into(),
         user_id.into(),
-        now.into(),
-        (now + SESSION_TTL_SECONDS).into(),
+        // D1's bind rejects a JS BigInt outright, and `worker`'s
+        // `From<i64> for JsValue` produces one -- cast through f64
+        // (lossless well under 2^53 for a unix-seconds timestamp),
+        // same fix already applied in h_approve_knock.
+        (now as f64).into(),
+        ((now + SESSION_TTL_SECONDS) as f64).into(),
     ])?
     .run()
     .await?;
@@ -260,8 +264,8 @@ pub async fn handle_register(mut req: Request, env: &Env) -> Result<Response> {
     .bind(&[
         hash.into(),
         email.clone().into(),
-        now.into(),
-        (now + VERIFICATION_TTL_SECONDS).into(),
+        (now as f64).into(),
+        ((now + VERIFICATION_TTL_SECONDS) as f64).into(),
     ])?
     .run()
     .await?;
@@ -337,7 +341,7 @@ pub async fn handle_verify(req: Request, env: &Env) -> Result<Response> {
         pending.email.clone().into(),
         password_hash.into(),
         salt.into(),
-        now.into(),
+        (now as f64).into(),
     ])?
     .run()
     .await?;
